Add files via upload
This program runs a beautfulsoup to extract NBA player statistics from nbareference.com. The stats are formatted into a pandas dataframe based on the year and the type of statistic. The five "save" functions use the getStats function to create local CSV files for each type of statistic within the range of years passed as paramaters. These are saved in folders created by the createFolder function within the directory variable.